backend/pyproject.toml
[project]
name = "startinsight-backend"
version = "0.1.0"
description = "StartInsight Backend - AI-powered business intelligence engine for startup ideas"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "StartInsight Team" }
]
dependencies = [
    # API Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.25",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "redis>=5.0.1",
    # AI & Agents (Phase 2 - included for completeness)
    "pydantic-ai>=0.0.13",
    "anthropic>=0.25.0",
    "openai>=1.12.0",
    # Web Scraping
    "firecrawl-py>=0.0.16",
    "praw>=7.7.1",
    "pytrends>=4.9.2",
    "tweepy>=4.14.0",  # Phase 7.1: Twitter/X integration
    # Task Queue
    "arq>=0.25.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",
    "apscheduler>=3.11.2",
    # Authentication
    "python-jose[cryptography]>=3.3.0",
    # SSE (Server-Sent Events) for admin real-time updates
    "sse-starlette>=1.6.5",
    # Phase 6: Payments & Email
    "stripe>=7.0.0",  # Phase 6.1: Payment integration
    "resend>=0.7.0",  # Phase 6.2: Email notifications
    # Phase 7: Rate Limiting
    "slowapi>=0.1.9",  # Phase 6.3: Rate limiting
    "prophet>=1.3.0",
    "rich>=14.2.0",
    "sentry-sdk[fastapi]>=2.0.0",
    "beautifulsoup4>=4.14.3",
    "bleach>=6.1.0",  # Input sanitization for XSS prevention
    "itsdangerous>=2.1.0",  # Token signing for email unsubscribe
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "aiosqlite>=0.19.0",
    "httpx>=0.26.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
